requirements.txt
feedparser==6.0.10
anthropic==0.38.0
fastapi==0.115.5
uvicorn==0.30.0
python-dotenv==1.0.0
apscheduler==3.10.4
requests==2.31.0
